:recycle: Move interfaces from class file
