+++ src/helpers/interfaces.ts
@@ -1,4 +1,34 @@
-import { Uppload } from "../uppload";
+import { Uppload, UpploadEffect, UpploadService } from "../";
+
+export type Elements = string | string[] | Element | Element[];
+export interface IUppload {
+  services: UpploadService[];
+  effects: UpploadEffect[];
+  isOpen: boolean;
+  error?: string;
+  activeService: string;
+  activeEffect: string;
+  settings: IUpploadSettings;
+  container: HTMLDivElement;
+}
+
+export interface IUpploadSettings {
+  value?: string;
+  bind?: Elements;
+  call?: Elements;
+  defaultService?: string;
+  lang?: { [index: string]: any };
+  uploader?: Uploader;
+  inline?: boolean;
+  customClass?: string;
+  multiple?: boolean;
+  compression?: number;
+  compressionMime?: string;
+  maxWidth?: number;
+  maxHeight?: number;
+  maxSize?: [number, number];
+  compressor?: (file: Blob) => Promise<Blob>;
+}
 
 export interface HandlersParams {
   upload: (file: Blob) => Promise<string>;
